[ADD] reports_cusomisations_tau Customise invoice, so, po reports and pos receipt [2484716]
