+++ reports_customisations_tau/views/pos_assets.xml
